요구사항 1차 2차 분리
